cbre_ui/backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
websockets
playwright
pinecone
openai
pandas
python-multipart
